+++ js/renderImageEffect.js
