+++ js/renderUploadMessage.js
@@ -1,5 +1,5 @@
 import { isEscapeKey } from './util.js';
-import { uploadPhotoSection, closeImageRedactor } from './uploadImage.js';
+import { uploadPhotoSection, closeImageRedactor } from './renderUploadImage.js';
 const uploadMessageContainer = document.querySelector('body');
 const successTemplate = document.querySelector('#success').content.querySelector('.success');
 const errorTemplate = document.querySelector('#error').content.querySelector('.error');
@@ -18,7 +18,7 @@ const closeErrorWindow = () => {
   uploadMessageContainer.removeEventListener('keydown', onDocKeydown);
 };
 
-function onDocKeydown (evt) {
+function onDocKeydown(evt) {
   const successWindow = document.querySelector('.success');
   const errorWindow = document.querySelector('.error');
   if (isEscapeKey(evt)) {
